# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Defines Canadian Insurance-Mortgage POC REST services"
  version: "1.0.0"
  title: "Canadian Insurance-Mortgage"
  contact:
    email: "kumari.shweta@hcl.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "@@host-details@@"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /findAllTransactionByType/{transactionType}:
    options:
      tags:
      - "Transaction"
      summary: "Get success for options"
      description: "This service used to send 200 for option method call"
      operationId: "findAllTransactionByTypeOptions"
      produces:
      - "application/json"
      parameters:
      - name: "transactionType"
        in: "path"
        description: "TransactionType either addLoan or addIPLetter"
        required: true
        type: "string" 
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "artWorkController"
    get:
      tags:
      - "Transaction"
      summary: "Get  information by transactionType"
      description: "This service is used to get  information by transactionType"
      operationId: "findAllTransactionByType"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "transactionType"
        in: "path"
        description: "TransactionType either addArtWork,updateArtWork or addConditionReport"
        required: true
        type: "string" 
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Data not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "artWorkController"
  /getTransactionByReferenceNum/{referenceNumber}:
    options:
      tags:
      - "Transaction"
      summary: "Get success for options"
      description: "This service used to send 200 for option method call"
      operationId: "getTransactionByReferenceNumOptions"
      produces:
      - "application/json"
      parameters:
      - name: "referenceNumber"
        in: "path"
        description: "Reference Number either artwork number or conditionreportnumber"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        400:
          description: "bad request"
        404:
          description: "not found"
      x-swagger-router-controller: "artWorkController"
    get:
      tags:
      - "Transaction"
      summary: "Get  information by transactionType"
      description: "This service is used to get  information by transactionType"
      operationId: "getTransactionByReferenceNum"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "referenceNumber"
        in: "path"
        description: "Reference Number either artwork number or conditionreportnumber"
        required: true
        type: "string" 
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Data not found"
        405:
          description: "Validation exception"
      x-swagger-router-controller: "artWorkController"
...
  /findByTransactionId/{transactionId}:
    options:
      tags:
      - "Transaction"
      summary: "Get  information by transactionId"
      description: "This service is used to get  information by transactionId"
      operationId: "getDetailsByTransactionIdOptions"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "transactionId"
        in: "path"
        description: ""
        required: true
        type: "string" 
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/transactiondata"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Audit not found"
      x-swagger-router-controller: "artWorkController" 
    get:
      tags:
      - "Transaction"
      summary: "Get  information by transactionId"
      description: "This service is used to get  information by transactionId"
      operationId: "getDetailsByTransactionId"
      security:
        - Bearer: []
      produces:
      - "application/json"
      parameters:
      - name: "token"
        in: "header"
        required: false
        type: "string"
      - name: "transactionId"
        in: "path"
        description: ""
        required: true
        type: "string" 
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/transactiondata"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Audit not found"
      x-swagger-router-controller: "artWorkController" 
securityDefinitions:
  dlpstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:ipnotice: "write IP notice"
      read:ipnotice: "read IP notice"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"  
definitions:
...
  transactiondata:
    type: "object"
    properties:
      transactionId :  
        type: "string"
      transactionType :
        type: "string"
      dateTime :
        type: "string"
      actor :
        type: "string"
      actorReference :
        type: "string"
      blockNumber :
        type: "string"
      additionalTags :
        type: "string"
...